#58 update mock generator
