+++ src/modules/common/db/mockData/index.ts
@@ -1,2 +1,3 @@
 export * from './users.mock'
 export * from './task.mock'
+export * from './createMockData'
