+++ src/modules/common/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './logger-options'
 export * from './transformer'
-export * from './createMockData'
